Tugas 5 revisi

+++ main/views.py
@@ -12,7 +12,6 @@ from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.http import HttpResponseForbidden
 
-...
 @login_required(login_url='/login')
 def show_main(request):
     filter_type = request.GET.get("filter", "all")  # default 'all'
@@ -29,37 +28,34 @@ def show_main(request):
         'products_list': products_list,
         'last_login': request.COOKIES.get('last_login', 'Never')
     }
-    return render(request, "main.html",context)
+    return render(request, "main.html", context)
 
 
+@login_required(login_url='/login')
 def create_products(request):
     form = ProductsForm(request.POST or None)
 
     if form.is_valid() and request.method == 'POST':
-        products_entry = form.save(commit = False)
+        products_entry = form.save(commit=False)
         products_entry.user = request.user
         products_entry.save()
         return redirect('main:show_main')
 
-    context = {
-        'form': form
-    }
-
+    context = {'form': form}
     return render(request, "create_products.html", context)
+
 
 @login_required(login_url='/login')
 def show_products(request, id):
     products = get_object_or_404(Products, pk=id)
-
-    context = {
-        'products': products
-    }
-
+    context = {'products': products}
     return render(request, "products_detail.html", context)
 
+
+@login_required(login_url='/login')
 def purchase_product(request, product_id):
-    product = get_object_or_404(Products, id=product_id)
-    
+    product = get_object_or_404(Products, pk=product_id)
+
     if request.method == 'POST':
         quantity = int(request.POST.get('quantity', 1))
         
@@ -69,32 +65,37 @@ def purchase_product(request, product_id):
         else:
             return render(request, 'out_of_stock.html')
 
+
 def show_xml(request):
     products_list = Products.objects.all()
     xml_data = serializers.serialize("xml", products_list)
     return HttpResponse(xml_data, content_type="application/xml")
+
 
 def show_json(request):
     products_list = Products.objects.all()
     json_data = serializers.serialize("json", products_list)
     return HttpResponse(json_data, content_type="application/json")
 
+
 def show_xml_by_id(request, products_id):
-   try:
-       products_item = Products.objects.filter(pk=products_id)
-       xml_data = serializers.serialize("xml", products_item)
-       return HttpResponse(xml_data, content_type="application/xml")
-   except Products.DoesNotExist:
-       return HttpResponse(status=404)
+    try:
+        products_item = Products.objects.filter(pk=products_id)
+        xml_data = serializers.serialize("xml", products_item)
+        return HttpResponse(xml_data, content_type="application/xml")
+    except Products.DoesNotExist:
+        return HttpResponse(status=404)
+
 
 def show_json_by_id(request, products_id):
-   try:
-       products_item = Products.objects.get(pk=products_id)
-       json_data = serializers.serialize("json", [products_item])
-       return HttpResponse(json_data, content_type="application/json")
-   except Products.DoesNotExist:
-       return HttpResponse(status=404)
-   
+    try:
+        products_item = Products.objects.get(pk=products_id)
+        json_data = serializers.serialize("json", [products_item])
+        return HttpResponse(json_data, content_type="application/json")
+    except Products.DoesNotExist:
+        return HttpResponse(status=404)
+
+
 def register(request):
     form = UserCreationForm()
 
@@ -104,8 +105,9 @@ def register(request):
             form.save()
             messages.success(request, 'Your account has been successfully created!')
             return redirect('main:login')
-    context = {'form':form}
+    context = {'form': form}
     return render(request, 'register.html', context)
+
 
 def login_user(request):
     if request.method == 'POST':
@@ -131,20 +133,29 @@ def logout_user(request):
     return response
 
 
+@login_required(login_url='/login')
 def delete_products(request, id):
     products = get_object_or_404(Products, pk=id)
+    
+    # hanya owner yang bisa hapus
+    if products.user != request.user:
+        return HttpResponseForbidden("You are not allowed to delete this product.")
+
     products.delete()
     return HttpResponseRedirect(reverse('main:show_main'))
 
+
+@login_required(login_url='/login')
 def edit_products(request, id):
     products = get_object_or_404(Products, pk=id)
+
+    if products.user != request.user:
+        return HttpResponseForbidden("You are not allowed to edit this product.")
+
     form = ProductsForm(request.POST or None, instance=products)
     if form.is_valid() and request.method == 'POST':
         form.save()
         return redirect('main:show_main')
 
-    context = {
-        'form': form
-    }
-
+    context = {'form': form}
     return render(request, "edit_products.html", context)
